backend/requirements.txt
# FastAPI backend
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
pydantic==2.11.7
pydantic_core==2.33.2
starlette==0.47.2
click==8.1.8
anyio==4.9.0
h11==0.16.0
idna==3.10
sniffio==1.3.1
typing-inspection==0.4.1
typing_extensions==4.14.1
annotated-types==0.7.0
exceptiongroup==1.3.0

# LangChain ecosystem (compatible versions)
langchain==0.1.16
langchain-core==0.1.50
langchain-community==0.0.35
langsmith==0.1.17
langchain-text-splitters==0.0.2

# Embeddings
huggingface-hub==0.23.0
sentence-transformers==2.7.0
transformers==4.41.1
torch==2.8.0

# Vector Store
qdrant-client==1.8.1

# Token counting
tiktoken==0.7.0

# Web search
requests==2.31.0

# .env config
python-dotenv==1.0.1
fastapi
qdrant_client
langchain_community
python-multipart
langchain_huggingface
langchain_qdrant